Fix merge conflicts with #46
