+++ app/(tabs)/settings.tsx
@@ -1,101 +1,121 @@
-import React, { useEffect, useState } from 'react';
-import { ActivityIndicator, StyleSheet, Text, View } from 'react-native';
-
-import { useRouter } from 'expo-router';
+import React, { useState, useEffect, useContext } from 'react';
 import { Ionicons } from '@expo/vector-icons';
-import { doc, getDoc } from 'firebase/firestore';
-
-import { Button } from '@/components';
-import { auth, db } from '@/services/firebase';
+import { View, Text, ActivityIndicator, TouchableOpacity} from 'react-native';
+import { StyleSheet } from 'react-native';
+import { doc, getDoc, getFirestore, updateDoc } from 'firebase/firestore';
+import { auth, db } from '@/app/logged-in/firebase';
+import { useRouter } from 'expo-router';
+import { getAuth } from 'firebase/auth';
 
 const Settings = () => {
-  const [isAdmin, setIsAdmin] = useState(false);
-  const [loading, setLoading] = useState(true);
-
+  const [adminStatus, setAdminStatus] = useState<boolean>(false); 
   const router = useRouter();
+
   const handleLogout = () => {
     router.push('/login')
   };
   const handleCreateAdmins = () => {
     router.push('/settings/create_admins');
   };
+  const createAnAccount = () => {
+    router.push('/create-account')
+  };
+
+  // Following function checks for admin status
   useEffect(() => {
-    const checkUserRole = async () => {
-      try {
-        // Get current user ID
-        const user = auth.currentUser;
-        if (!user) {
-          console.error('User not logged in');
-          setLoading(false);
-          return;
-        }
-
-        // Fetch user role from Firestore
-        const userDoc = await getDoc(doc(db, 'users', user.uid));
-        if (userDoc.exists()) {
-          const { role } = userDoc.data();
-          setIsAdmin(role === 1 || role === 2); // Admin roles are 1 or 2
-        } else {
-          console.error('User document does not exist');
-        }
-      } catch (error) {
-        console.error('Error fetching user role:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    checkUserRole();
+    setUserRole();
   }, []);
+  const setUserRole = async () => {
+    const auth = getAuth();
+    const user = auth.currentUser;
+    if (user) {
+      const userDoc = await getDoc(doc(db, "users", user.uid));
+      if (userDoc.exists()) {
+        const userRole = userDoc.data().role;
+        setAdminStatus(userRole === 1 || userRole === 2);
+      } else {
+        console.log("No user document found!");
+      }
+    }
+  };
 
-  if (loading) {
-    return (
-      <View style={styles.screen}>
-        <ActivityIndicator size="large" color="#000" />
-      </View>
-    );
-  }
   return (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Button style={styles.logoutButton} onPress={handleLogout}>
-        <Ionicons name="log-out-outline" size={25} color="#fff" />
-      </Button>
-      <Text>Setting Screen</Text>
-      {isAdmin && (<>
-        <Button onPress={handleCreateAdmins}>
-          Make someone else an admin
-        </Button>
-        <Ionicons
-          name="lock-closed"
-          size={24}
-          color="black"
-          style={styles.lockIcon}
-        />
-      </>)}
+        <TouchableOpacity style={styles.logoutButton} onPress={handleLogout}>
+            <Ionicons name="log-out-outline" size={25} color="#fff" />
+        </TouchableOpacity>
+        <Text>Setting Screen</Text>
+        {adminStatus && (<>
+          <TouchableOpacity style={styles.button} onPress={handleCreateAdmins}>
+            <Text style={styles.buttonText}>Make someone else an admin</Text>
+          </TouchableOpacity>
+          <Ionicons
+            name="lock-closed"
+            size={24}
+            color="black"
+            style={styles.lockIcon}
+          />
+        </>)}
     </View>
   );
-};
+}
 
 export default Settings;
 
 const styles = StyleSheet.create({
-  screen: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  lockIcon: {
-    marginRight: 15, // Adjust as needed for positioning
-  },
-  logoutButton: {
-    position: 'absolute',
-    top: 10,
-    left: 20,
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '#333',
-    padding: 5,
-    borderRadius: 5,
-    zIndex: 10, // Ensure the logout button is on top
-  },
-});
+    screen: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center',
+      },
+      lockIcon: {
+        marginRight: 15, // Adjust as needed for positioning
+      },
+    container: {
+      flex: 1,
+      justifyContent: 'flex-end', // Tab navigator will sit at the bottom
+    },
+    screenContainer: {
+      flex: 1,
+      top: -50,
+      justifyContent: 'center',
+      alignItems: 'center',
+    },
+    logoutButton: {
+      position: 'absolute',
+      top: 10,
+      left: 20,
+      flexDirection: 'row',
+      alignItems: 'center',
+      backgroundColor: '#333',
+      padding: 5,
+      borderRadius: 5,
+      zIndex: 10, // Ensure the logout button is on top
+    },
+    logoutText: {
+      color: '#fff',
+      marginLeft: 5,
+    },
+    tabs: {
+      flex: 1,
+      justifyContent: 'flex-end',
+      bottom: 0,
+      backgroundColor: '#333',
+    },
+    button: {
+      backgroundColor: '#333',
+      padding: 12,
+      borderRadius: 5,
+      alignItems: 'center',
+      margin: 5,
+      display: 'flex',
+      justifyContent: 'center',
+    },
+    buttonText: {
+      color: '#fff',
+      fontSize: 12,
+      fontWeight: 'bold',
+      textAlign: 'center',
+  
+    },
+  });
